services/web/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Werkzeug==2.3.7
gunicorn==20.1.0
psycopg2-binary==2.9.6
